packages/scoring/pyproject.toml
[project]
name = "miami-scoring"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "miami-common",
    "miami-domain",
    "miami-features",
    "polars>=1.18",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/miami_scoring"]
